// src/app/web/api/tranUser.ts
export interface TranUser {
    userId: number;
    orgId: number;
    divId: number;
    locationId: number;
    userLocId: number;
    positionId: number;
    positionName: string;
    orgName: string;
    divName: string;
    remark: string;
    usercode: string;
    userRole: string;
    userType: string;
    firstnameEn: any;
    lastnameEn: any;
    ascCategory: string;
    ascPosition: string;
    prefixName: string;
    prefixNameLovName : string
    firstname: string;
    lastname: string;
    email: string;
    userStatus: string;
    userAsccode: string;
    period: string;
    effectiveStartDate: string;
    effectiveEndDate: any;
    effectiveStartDateStr: any;
    effectiveEndDateStr: any;
    mobile: string;
    phone: string;
    createdBy: string;
    updatedBy: string;
    createdDate: string;
    updatedDate: string;
    username: string;
    userTypeName: string;
    userPeriodName: string;
    ascCategoryName: string;
    ascPositionName: string;
}